Fix print function for non-string args
